feature: add state variable for tracking if the content is on a mobile device, use the new state in components
