+++ src/components/experience/experience-selector-list.tsx
@@ -2,30 +2,17 @@ import { useDispatch, useSelector } from "react-redux";
 import ExperienceSelectorButton from "./experience-selector-button";
 import { RootState } from "../../store/store";
 import { experienceActions } from "../../store/experience";
-import { useEffect, useState } from "react";
 
 interface Props {
   experiences: any[];
 }
 
 const ExperienceSelectorList = (props: Props) => {
-  const [screenWidth, setScreenWidth] = useState(window.innerWidth);
+  const isMobile = useSelector((state: RootState) => state.ui.isMobile);
   const dispatcher = useDispatch();
   const activeExperienceId = useSelector(
     (state: RootState) => state.experience.activeExperienceId
   );
-
-  useEffect(() => {
-    function handleResize() {
-      setScreenWidth(window.innerWidth);
-    }
-
-    window.addEventListener("resize", handleResize);
-
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
 
   const setActiveExperienceIdHandler = (id: number) => {
     dispatcher(experienceActions.setActiveExperienceId(id));
@@ -34,7 +21,10 @@ const ExperienceSelectorList = (props: Props) => {
   const experienceSelectorButtons = props.experiences.map(
     (experience, index) => {
       return (
-        <li key={index} className="bg-blue-800 bg-opacity-60 border-solid border-x border-blue-300 sm:border-none sm:bg-transparent">
+        <li
+          key={index}
+          className="bg-blue-800 bg-opacity-60 border-solid border-x border-blue-300 sm:border-none sm:bg-transparent"
+        >
           <ExperienceSelectorButton
             id={index}
             experienceName={experience.experienceName}
@@ -47,7 +37,7 @@ const ExperienceSelectorList = (props: Props) => {
 
   return (
     <div className="relative grid w-full grid-cols-1 grid-rows-1 sm:row-start-1 sm:row-end-2 sm:inline sm:grid-rows-none sm:grid-cols-none">
-      {screenWidth > 640 && (
+      {!isMobile && (
         <div
           className={`absolute top-0 left-0 w-1 h-10 rounded-3xl bg-emerald-300
                       transition-all duration-150`}
